+++ framework/domain/src/main/java/is/codion/framework/domain/entity/condition/CustomCondition.java
@@ -24,7 +24,7 @@ import java.util.List;
  * A custom {@link Condition} based on a {@link ConditionString}.
  * <p>
  * Custom conditions are used to create query conditions that cannot be created with the
- * standard {@link Condition}, {@link ColumnConditionFactory} or {@link ForeignKeyConditionFactory} APIs.
+ * standard {@link Condition}, {@link ColumnConditions} or {@link ForeignKeyConditions} APIs.
  * They enable complex SQL constructs such as subqueries, JOINs, window functions, and
  * database-specific functionality.
  * <p>