+++ framework/domain/src/main/java/module-info.java
@@ -30,8 +30,8 @@
  * <li>{@link is.codion.framework.domain.entity.attribute.ForeignKeyDefinition}
  * <li>{@link is.codion.framework.domain.entity.condition.Condition}
  * <li>{@link is.codion.framework.domain.entity.condition.ColumnCondition}
- * <li>{@link is.codion.framework.domain.entity.condition.ColumnConditionFactory}
- * <li>{@link is.codion.framework.domain.entity.condition.ForeignKeyConditionFactory}
+ * <li>{@link is.codion.framework.domain.entity.condition.ColumnConditions}
+ * <li>{@link is.codion.framework.domain.entity.condition.ForeignKeyConditions}
  * </ul>
  * @uses is.codion.framework.domain.Domain
  */